core: Allow optional dependencies.
Allow modules to declare themselves as optional through the use of
the `IS_OPTIONAL` attribute.

Modules that declare themselves as optional MUST also provide a
`installed()` function that consuming modules can use to gate
execution.

Optional modules, will be considered in the dependency graph
regardless of package installation status to ensure correct
execution order.

It is up to the consumer of a optional module to gate execution
on actual availability at runtime.

Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>