main: Introduce test concurrency argument.
The commit in the fixes tag hard coded tempest to run tests
serially.  While this may be required in constrained
environments, it prolongs execution unnecessarily when more
resources are available.

Introduce a --concurrency argument to the test sub-command which
defaults to 1.  When providing the 'auto' keyword, concurrency is
set to the rounded up value of number of CPUs divided by 3.

Fixes: 68aad51e3732 ("Use ubuntu image to ensure multi-platform")
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>